+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for NetBSD VAX via the SIMH simulator using packer-plugin-simh
 
 ## [0.6.0] - 2026-04-29
 ### Changed